+++ src/components/icon/icon.style.js
@@ -47,6 +47,7 @@ const StyledIcon = styled.span`
     type,
     fontSize,
     disabled,
+    hasTooltip,
   }) => {
     let finalColor;
     let finalHoverColor;
@@ -128,6 +129,13 @@ const StyledIcon = styled.span`
         
         display: block;
       }
+
+      ${hasTooltip &&
+      `
+        :focus {
+          outline: 2px solid ${theme.colors.focus};
+        }
+      `}
 
       ${margin}
     `;